% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLi.R
\name{checkInput}
\alias{checkInput}
\title{checkInput}
\usage{
checkInput(corr, mod, N_sample, abs, pprob, saveAs)
}
\arguments{
\item{corr}{Correlation matrix}

\item{mod}{Model file (landmarks classified into modules)}

\item{abs}{Logical - whether to use abosulte values of correlations}

\item{pprob}{posterior probability cut off}

\item{saveAs}{filename to save output too}
}
\description{
Checks the input of EMMLi to make sure it's all what is expected. Throws errors with
a descriptive error when it is not.
}
\keyword{internal}
