% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleSummary.R
\name{getBestMods}
\alias{getBestMods}
\title{getBestMods}
\usage{
getBestMods(rs)
}
\arguments{
\item{rs}{The output of subsampleEMMLi}
}
\description{
Collates the best fitting models from each of the analyses in a subsampledEMMLi analysis
}
\keyword{internal}
