% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleEMMLi.R
\name{getCorrs}
\alias{getCorrs}
\title{getCorrs}
\usage{
getCorrs(emm, models, corr)
}
\arguments{
\item{emm}{A fitted EMMLi model (output of EMMLi).}

\item{models}{A data frame defining the models. The first column should contain the landmark names
as factor or character. Subsequent columns should define which landmarks are contained within each
module with integers, factors or characters. If a landmark should be ignored for a specific model
(i.e., it is unintegrated in any module), the element should be NA.}

\item{corr}{The original correlation matrix used in the EMMLi analysis that generated
the input.}
}
\value{
The correlations within and between modules of the best model.
}
\description{
This function takes a fitted EMMLi model and the models used to fit it, and then returns
all the correaltions for the best fitting model. This is recycled from the main EMMLi
function - Possible that this will change with the refactoring of EMMLi.
}
\keyword{internal}
