% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLi.R
\name{getModules}
\alias{getModules}
\title{getModules}
\usage{
getModules(varlist, symmet, corr, mod, corr_list)
}
\arguments{
\item{varlist}{internal variable from EMMLi}

\item{symmet}{internal variable from EMMLi}

\item{corr}{correlation matrix}

\item{mod}{Landmark classifications}

\item{corr_list}{internal variable from EMMLi}
}
\description{
Breaks correlation matrix up into discrete modules (within and between) based on
a model table (landmarks classified into modules)
}
\keyword{internal}
