% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleSummary.R
\name{getRos}
\alias{getRos}
\alias{getRhos}
\title{getRhos}
\usage{
getRhos(rs)
}
\arguments{
\item{rs}{The output of subsampleEMMLi}
}
\description{
Collates the rhos for each of the analyses in a subsampleEMMLi output. Collects them
together based on the model that derived them, like with like.
}
\keyword{internal}
