% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRSAL.R
\name{IRSAL}
\alias{IRSAL}
\title{IRSAL}
\usage{
IRSAL(atlas, landmarks, initial_fixed, n, reps, write.out = FALSE,
  sp = NULL, ...)
}
\arguments{
\item{atlas}{object of class "atlas" created by \link[Morpho]{createAtlas}}

\item{landmarks}{k x 3 x n array containing reference landmarks of the sample
or a matrix in case of only one target specimen.}

\item{initial_fixed}{The fixed points for the initial
\link[Morpho]{placePatch} analysis.}

\item{n}{The number of patched points to use in each iteration of IRSAL.}

\item{reps}{The number of times to repeat the IRSAL procedure.}

\item{write.out}{If TRUE then the patched points from each iteration, for
each species, will be written to the working directory. This will NOT include
the temporary anchor points.}

\item{sp}{If write.out is true, then this determines which species is
written out. Defaults to 1 (the first species in the dataset). 2 would be
the second species in the dataset etc.}

\item{...}{Additonal arguments passed to \link[Morpho]{placePatch}}
}
\description{
IRSAL
}
