% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMLi.R
\name{moduleLikelihoods}
\alias{moduleLikelihoods}
\title{moduleLikelihoods}
\usage{
moduleLikelihoods(all_modules, abs, N_sample, correction = "normal")
}
\arguments{
\item{all_modules}{output from getModules - internal to EMMLi}

\item{abs}{use absolute values or not}

\item{N_sample}{sample size}

\item{correction}{Experimental - whether to calculate number of parameters normally,
or number of parameters + number of modules - 1}
}
\description{
Calculates the maximum likelihood value of rho for each module (between/within) from
getModules
}
\keyword{internal}
