% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotMeanNetwork}
\alias{plotMeanNetwork}
\title{plotMeanNetwork}
\usage{
plotMeanNetwork(summary, ...)
}
\arguments{
\item{summary}{The output of \link[EMMLiv2]{summariseResults}}

\item{...}{Additional arguments for plotNetwork (see \link[EMMLiv2]{plotNetwork})}
}
\description{
Plots the mean network inferred from the rhos returned from a set of subsampled analyses.
}
