% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{plotNetwork}
\usage{
plotNetwork(rhos, module_names = NULL, linecolour = "#56B4E9",
  title = NULL, layout = NULL)
}
\arguments{
\item{rhos}{The rhos that come out of an \link[EMMLiv2]{EMMLi} analysis}

\item{module_names}{The names of the modules - if absent generic numbers are used.
If "rhos" then the nodes are named with the within-module rho.}

\item{linecolour}{The colour of the joining lines. If a single colour, then lines will
be that colour with width and transparency adjusted to the strength of the correlation.
If "viridis" then lines are coloured according to the viridis colour palette, with darker
colours corresponding to stronger correlations.}

\item{title}{Title for the plot.}

\item{layout}{A matrix describing the positions of each module on the canvas.
See \link[qgraph]{qgraph} for details.}
}
\value{
A plotted network of the relationships between and within modules.
}
\description{
Plots a network diagram of the output of an EMMLi analysis. Nodes are proportionally
sized to within-module rho, and lines are larger and darker according to between-module
rho.
}
\examples{
data("macacaCorrel")
data("macacaModels")
emm <- EMMLi(corr = macacaCorrel, mod = macacaModels, N_sample = 20)
plotNetwork(emm$rho[[1]], linecolour = "viridis")
}
