% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotRandomSubsamples}
\alias{plotRandomSubsamples}
\title{plotRandomSubsamples}
\usage{
plotRandomSubsamples(subsamples, n, ...)
}
\arguments{
\item{n}{The number of random subsamples to plot.}

\item{...}{Additional arguments for plotNetwork (see \link[EMMLiv2]{plotNetwork})}

\item{subsasmples}{The output of the execution of \link[EMMLiv2]{subSampleEMMLi}}
}
\value{
A gridded plot with a network plot for each of n random subsamples.
}
\description{
Generates network plots for a given number of randomly selected subsampled EMMLi analyses,
as returned by subSampleEmmli.
}
\examples{
ssemm <- subsampleEmmli(landmarks = landmarks, models = models, fractions = 0.4, min_landmark = 5, nsim = 25)
plotRandomSubsamples(subsamples = ssemm, n = 9, linecolour = "viridis")
}
