% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleEMMLi.R
\name{subSampleEmmli}
\alias{subSampleEmmli}
\alias{subSampleEMMLi}
\title{subSampleEMMLi
This takes some landmarks, and then will subsample them down to one or more subsampling
fractions, calculate the correaltion matrix (using dotcorr only) and then fit EMMLi,
returning the result of all the subsampling (if multiple fractions). If a single fraction
the nsim argument is required with a number of times to repeat subsampling at that level.}
\usage{
subSampleEMMLi(landmarks, fractions, models, min_landmark, aic_cut = 7,
  return_corr = TRUE, nsim = NULL)
}
\arguments{
\item{landmarks}{Landmarks dataframe}

\item{fractions}{Either a single subsampling fraction (in which case nsim is required) or a vector of fractions.}

\item{models}{The models to test.}

\item{min_landmark}{The minimum number of landmarks to subsample to}

\item{aic_cut}{This is the threshold of dAICc below which two models are considered to be not different. When this occurs multiple models can be returned as the best.}

\item{return_corr}{Logical - if TRUE then the full correlations of within and between modules are returned for the best models.}

\item{nsim}{If a single subsampling fraction, this is the number of times that subsampling fraction is repeated.}
}
\description{
subSampleEMMLi
This takes some landmarks, and then will subsample them down to one or more subsampling
fractions, calculate the correaltion matrix (using dotcorr only) and then fit EMMLi,
returning the result of all the subsampling (if multiple fractions). If a single fraction
the nsim argument is required with a number of times to repeat subsampling at that level.
}
