% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleEMMLi.R
\name{subsampleLandmarks}
\alias{subsampleLandmarks}
\title{subsampleLandmarks}
\usage{
subsampleLandmarks(landmarks, fraction, models, min_landmark)
}
\arguments{
\item{landmarks}{A 2D dataframe of landmarks}

\item{fraction}{The decimal fraction to subsample down to. e.g. 0.2 will return 20% of the original landmarks.}

\item{models}{The models for the landmarks}

\item{min_landmarks}{The minimum number of landmarks for a module in any model. When subsampling goes below this threshold landmarks are randomly seleclted from the original data and added back in in order to make up to min_landmarks. This means that sometimes the exact fraction desired is not returned.}
}
\description{
This function randomly removes landmarks from a dataset, and adjusts the corresponding
models object to match the newly subsampled landmark data. Internal, called by
subsampleEMMLi.
}
\keyword{internal}
