% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleEMMLi.R
\name{subsampleLandmarks}
\alias{subsampleLandmarks}
\title{subsampleLandmarks}
\usage{
subsampleLandmarks(landmarks, fraction, models, min_landmark)
}
\arguments{
\item{landmarks}{A 2D array of xyz landmarks to subsample from. Will be turned into a correlation
matrix using \link[paleopmorph]{dotcorr} for EMMLi analysis after subsampling.}

\item{fraction}{The decimal fraction to subsample down to. e.g. 0.2 will return 20% of the
original landmarks.}

\item{models}{A data frame defining the models. The first column should contain the landmark names
as factor or character. Subsequent columns should define which landmarks are contained within each
module with integers, factors or characters. If a landmark should be ignored for a specific model
(i.e., it is unintegrated in any module), the element should be NA.}

\item{min_landmark}{The minimum number of landmarks to subsample to. This ensures that a module isn't
totally removed during random subsampling. When subsampling causes a landmark to be removed or to be
subsampled below this threshold landmarks are drawn from the original module at random and added back in.
This means that sometimes (especially with low subsampling fractions and/or the presence of small
modules in a model) the actual subsampling level is higher than the requested subsampling. In these
cases a warning is printed to the screen.}
}
\description{
This function randomly removes landmarks from a dataset, and adjusts the corresponding
models object to match the newly subsampled landmark data. Internal, called by
subsampleEMMLi.
}
\keyword{internal}
