% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleSummary.R
\name{subsampleSummary}
\alias{subsampleSummary}
\title{subsampleSummary}
\usage{
subsampleSummary(subsamples)
}
\arguments{
\item{subsamples}{The output of subsampleEMMLi}
}
\description{
Summarises the output of subsampleEMMLi to allow easier comparison between multiple
subsampled analyses
}
